moved urls to different file
