+++ Frontend/src/Components/constant/api.jsx
@@ -1,0 +1,2 @@
+export const EXCUSE_URL = "http://localhost:3000/excuse";
+export const USERS_URL = "http://localhost:3000/users";
